% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Say goodbye to someone}
\usage{
goodbye(name)
}
\arguments{
\item{name}{The name of the person to say goodbye to}
}
\value{
A string saying goodbye to the person, reusing \code{\link[=hello]{hello()}} as the prefix
}
\description{
Say goodbye to someone
}
\examples{
goodbye("world")
}
