% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to someone}
\usage{
hello(name)
}
\arguments{
\item{name}{The name of the person to say hello to}
}
\value{
A string saying hello to the person
}
\description{
Say hello to someone
}
\examples{
hello("world")
}
