% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_goodbye.R
\name{hello_goodbye}
\alias{hello_goodbye}
\title{Say hello and goodbye to someone}
\usage{
hello_goodbye(name)
}
\arguments{
\item{name}{The name of the person}
}
\value{
A string saying hello and goodbye to the person
}
\description{
Say hello and goodbye to someone
}
\examples{
hello_goodbye("world")
}
