% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellogoodbye.R
\name{hellogoodbye}
\alias{hellogoodbye}
\title{Say hello and goodbye to someone}
\usage{
hellogoodbye(name)
}
\arguments{
\item{name}{The name of the person}
}
\value{
A string saying hello and goodbye to the person
}
\description{
Say hello and goodbye to someone
}
\examples{
hellogoodbye("world")
}
